Fixes by Brian Walenz on trunk:
Fix for crash and memory corruption in unitigger.

Analyze_Overlap_AS() -- REMOVED.
Analyze_Affine_Overlap_AS() -- FIXED, interface changed.
Analyze_Affine_Overlap_IndelSize_AS() -- REMOVED.

All these were computing and returning a list of the positions in an alignment at which errors of some type (mismatch, etc) occur.  This was both never used, and buggy -- an array out of bounds.  The bug was causing crashes on x86_64 on large assemblies, and who knows what on alpha.  Mostly because this functionality was never used, but partially because fixing properly would introduce an ugly interface, the functionality was removed.

NOTE: Analyze_Affine_Overlap_AS() now has an additional parameter, the largest block of mismatch, which can be set to NULL if you don't care about it.  This functionality used to be in Analyze_Affine_Overlap_IndelSize_AS(), which has been removed.

NOTE: Analyze_Overlap_AS() was NEVER used, so it is now disabled.