Change that may increase robustness of MultiAlignContig() to difficult-to-find overlaps among unitigs.
The change fixes an apparent bug in the setting of try_contained, so that while we still initially try only non-contained unitigs as the afrag, if those fail, we can go back and try also contained unitigs.  (N.B. the containment status of a unitig can be updated during MultiAlignContig, if the alignment indicates a containment that was not implied by the ICM itself.)  This may help robustness in the case where unitigs A, B, and C are like so:

        A --------------
          B  --------
               C -------------

where B and C can be made to overlap but A and C cannot.  One might argue about whether such cases should be rescued ... but making such a call is not something consensus is designed to do and we typically prefer it to simply succeed on the given layout.