1) Add more clear ranges.  Both OBT and ECR didn't have enough ranges to store all the useful info, and the old OVL and CNS ranges were never used.  Now there are 10 ranges, three are read-only from the frag input, three are currently unused (OBT got two, ECR got two).
2) Replace the ReadStruct with the fragRecord.  Basically the same
idea, slightly improved.  Most functions change syntax from
getVal(&val) to val = getVal().  It is smarter about knowing if
sequence, quality, etc, has been read in, and will fail if one asks
for sequence, but didn't ask to have it read in.

This (and also storing this in the frag itself) gets rid of LOTS of
sequence accesses to just get the length.  In particular, eCR used to
read in the sequence, just to find the length of the read.  It can now
immediately get the length without that step.  Was used EXTENSIVELY
when finding a frag that could be extended.

3) safeRead() and safeWrite() replaced by AS_UTL_safeRead() and
AS_UTL_safeWrite().  A NOP, except for handling of eof on reading,
which looks like a NOP, but I'm not sure.

4) Slight optimization of the frag stream.  On creation, you must now
specify what is to be streamed.  If only the INF, then no seq or qlt
will be read.  Now use the streaming versions for reading seq and qlt.

5) The 'gatekeeper' executable now also dumps the store.  It replaces
dumpFragStoreAsFasta, dumpFragStoreAsOFG, make_OFG_from_FragStore,
dumpFragStore, etc, etc.