Hooray!  Partitioned stores exist again!
The interface to load a partition into memory is:
  gkpStore = openGateKeeperStore(whatever)
  loadGateKeeperPartition(gkpStore, partition_number)

There is no change in how a fragment is accessed.  getFrag() is smart
enough to know that the store is partitioned, and it will get the
fragment from the (in memory) partition.  If the fragment is not
in the partition, gatekeeper could fallback and load it from the full
store (on disk), but currently doesn't.  It just fails.