The thread locking mutexes were too fine-grained, resulting in lots of overhead.  A few mutexes could simply be removed, and were documented as such.
For partial overlaps, keep a buffer of overlaps found, and when the
buffer is full, write them to an AS_OVS binary overlap file.  The
buffering doesn't need a mutex, while the writing does.  The previous
ascii output is no longer supported.

Normal overlaps can still be reported as Overlap messages if the -P
option is used, otherwise they too are buffered and written to an
AS_OVS binary overlap file.  The single line format is no longer
supported -- but you can get that by converting the binary to ascii.

Finally:
  FOR_CARL_FOSLER has been removed.
  Contig mode and SHOW_SNPS might not be thread safe.
  AS_OVL_overlap_common.h has lots of whitespace changes.