Fix internal CGW distances.
Previously, CGW had two definitions of clone distance; (mean, stddev)
and (mu, sigma).  I suspect that one of those was intended to be
scratch space (mu, sigma) for calculating an updated distance to be
placed into (mean, stddev).  Unfortunately, both of those were used
throughout CGW, and they were not always the same.

Now, there is only one distance: (mu, sigma).  If CGW doesn't start
from a checkpoint, the distance is loaded from gatekeeper, and
reestimated very soon using unitigs.  If started from a checkpoint,
the values from the checkpoint are used.

runCA-OBT no longer needs to compute 6-clonesize, and does not by
default.

THIS COMMIT BREAKS CHECKPOINT COMPATIBILITY!  Earlier checkpoints
cannot be used.

Some details:

1) cgw now creates files stat/*.dst containing proto-io messages that
can be used to update the fragstore with the correct distances.  Three
files exist, one for unitigs, contigs and scaffolds.  The scaffold and
contig estimates should be fed to gatekeeper in that order.

2) The CIFragT in InputDataTypes_CGW.h changed.  mateStatus was
removed -- it was never used.  Its type was MateStatusType, which
narrowly missed MateStatType defined in AS_MSG_pmesg.h, leading to
confusion.  Note that there is a mateStatus in Instrument_CGW.h; not
the same thing.

3) The DistT in InputDataTypes_CGW.h changed, removing mean, stddev
and a temporary VA.  Note the ugly hack needed to read this in from
disk, to clear out a pointer.

4) The constant UNASSIGNED is now UNASSIGNED_MATE.