Yet another merged commit caused by a bizarre FreeBSD-specific IO bug.
1) AS_UTL_fileIO: Replace the CDS_FSEEK and CDS_FTELL macros with real
functions.  They are much safer (in that they actually check return
values), can optionally be debugged, and AS_UTL_fseek() checks if it
can skip the seek because the fp is already there....EXCEPT on
FreeBSD, which seems to occasionally (reliably when it occurs) screw
up the fp on fseek/fwrite for some frg files.

2) AS_PER: Muck around in AS_PER_genericStore.[ch]: Simplify the
creation of a memory store from a file store.  And other mucking when
I thought the FreeBSD problem was here.

3) AS_GKP: Add the "-E errorfile" option to gatekeeper to write
messages that do not pass to 'errorfile', along with comments as to
why they failed.  Update formatting on error messages.