1) Remove the -Q and -e options from gatekeeper.  Checking quality is a fairly pointless act now, and we've been stung enough times when we forgot to use -Q.  See #2, #3, #4 and #5 for -e.
2) Make gatekeeper much more forgiving on errors that can be trivially
corrected.

3) Make gatekeeper load ALL POSSIBLE fragments, even if this results
in them being immediately deleted.  This is for accounting.

4) gatekeeper -dumpinfo reports lots of stats about what was loaded,
and what currently exists in the store.

5) gatekeeper will fail (return an error code) iff a FRG references an
unknown library.  This error indicates either a library record didn't
load properly, or your input frags might have a serious problem, like
getting the UID for the library wrong.

This commit breaks the gkpStore/gkp file, but it is trivial to update an
old store by appending a bunch of zero bytes to the end of that file.