Refactor the chunk graph.  Contigs MUCH better now.
Two significant bugs found, but they're buried in heavy refactiong.
Line #'s are in the previous version of AS_BOG_ChunkGraph.cc.

Line 137 -- If the path length was known, the old code would return a
length of 1.  We now return the path length.

Line 151 -- If the path ended because we ran off the end of the path
(so, not a cycle, and we never ran into an existing path) the path
lengths would NOT be updated to be the length from the END.  Instead,
path lengths would be the length from the START.