Tiny bit of dead code removal (CheckForContainmentContigs(), and the *_CONTIGGING defines).
In InsertScaffoldContentsIntoScaffold(), scaffold lengths are adjusted
or corrected before and after the insert, based on the actual contig
positioning.  This fixes a crash where the scaffold was shorter than
the contig positioning claimed.  The root cause of this was not discovered.

Both InsertScaffoldContentsIntoScaffold() and InsertCIInScaffold() were
slightly cleaned up.

A handful of compiler warnings were fixed.