Replace the per node linked-list of edges with a balanced-tree of edges.
Checkpoints are still compatible, but these members are obsolete and eventually
need to be removed:
  in NodeCGW_T:  edgeHead
  in EdgeCGW_T:  nextALink, prevALink, nextBLink, prevBLink

CreateInitialContigEdges() was refactored into the existing BuildInitialContigs().

Scaffold edge loading from a checkpoint is disabled.  This impacts only caqc items
TotalScaffoldLinks and MeanScaffoldLinkWeight.

MergeEdges_CGW.c is now using the STL sort, instead of the C sort.