iAdd option 'gridEngine' to select either SGE or LSF.
Unset SGE_TASK_ID in runCA.  If it is set, overlapper and consensus scheduling
are confused.  This was only an issue if runCA was run from within a task array.

Use JOB_ID instead of SGE_TASK_ID to decide if runCA is being run locally
(scriptOnGrid=0) or under SGE control (scriptOnGrid=1).